// deploy/restpe-contabilidad-front-end/src/app/modules/almacen/domain/models/gestion-devolucion.entity.ts
export interface GestionDevolucionEntity {
  almacen_codigo: string;
  facturaA: string;
  fechaR: string;
  gestion_devolucion_producto: string;
  gestion_devolucion_proveedor: string;
  gestion_devolucion_almacen: string;
  devolucion: string;
  gestion_devolucion_responsable: string;
  gestion_devolucion_estado: string;
}