// deploy/restpe-contabilidad-front-end/src/app/modules/almacen/domain/models/registro-perdidas.entity.ts
export interface RegistroPerdidasEntity {
  almacen_codigo: string;
  fechaR: string;
  codproducto: string;
  registro_perdidas_producto: string;
  registro_perdidas_almacen: string;
  tipo: string;
  cantidadP: number;
  registro_perdidas_medida: string;
  valorU: string;
  valorT: string;
  registro_perdidas_responsable: string;
  registro_perdidas_estado: string;
}